custom_chat/requirements.txt
openai
langchain
streamlit
streamlit-chat
python-dotenv
tiktoken
